Implemented google login
1. Impelemented google login and logout
2. Created a users(members) table to store the authentication info for each account
3. Added some constraints for the ability to login to some pages
4. Made the update/delete functionality available only to the user who created the movie
5. Made the new movie functionality available only for authorized users

+++ movie_catalog.py
@@ -1,8 +1,22 @@
+import httplib2
+import json
+import random
+import requests
+import string
 import utils
-from flask import render_template, redirect, request, url_for, flash
+from flask import (render_template, redirect, request,
+                   url_for, flash, session as login_session,
+                   abort, make_response)
+from flask_login import login_user, logout_user, current_user, login_required
 from forms import NewMovieForm, UpdateMovieForm
-from models import db, Movie, Genre
+from models import db, Movie, Genre, Member
+from oauth2client.client import flow_from_clientsecrets, FlowExchangeError
 from setup import app
+from sqlalchemy.orm.exc import NoResultFound
+
+
+CLIENT_ID = json.loads(
+    open('client_secret.json', 'r').read())['web']['client_id']
 
 
 @app.route('/')
@@ -43,6 +57,7 @@ def movie(genre_id, movie_id):
 
 
 @app.route('/movie/new', methods=['GET', 'POST'])
+@login_required
 def create_movie():
     '''Create a new movie and commit it to the database.'''
 
@@ -62,13 +77,16 @@ def create_movie():
                       duration_hrs=form.duration_hrs.data,
                       duration_mins=form.duration_mins.data,
                       rate=form.rate.data,
-                      storyline=form.storyline.data)
+                      storyline=form.storyline.data,
+                      member=current_user)
+
         # check if the user uploaded an image, if not use the default.
         if request.files.get('poster'):
             uploaded_poster_img = form.poster.data
             stored_poster_fn = utils.save_image(uploaded_poster_img,
                                                 'static/imgs/movie_pics')
             movie.poster = stored_poster_fn
+
         db.session.add(movie)
         db.session.commit()
         flash(f'You have successfully added the movie "{movie.title}"'
@@ -80,10 +98,15 @@ def create_movie():
 
 @app.route('/genre/<int:genre_id>/movie/<int:movie_id>/update',
            methods=['GET', 'POST'])
+@login_required
 def update_movie(movie_id, genre_id):
     '''Update Movie Information and commit to the database.'''
 
     movie = Movie.query.get(int(movie_id))
+
+    if movie.author != current_user:
+        abort(403)
+
     form = UpdateMovieForm(obj=movie)
     form.genre.choices = [(movie.genre.name, movie.genre.name)]
 
@@ -94,12 +117,14 @@ def update_movie(movie_id, genre_id):
         movie.duration_mins = form.duration_mins.data
         movie.rate = form.rate.data
         movie.storyline = form.storyline.data
+
         # check if the user uploaded an image, if not use the default.
         if request.files.get('poster'):
             uploaded_poster_img = form.poster.data
             stored_poster_fn = utils.save_image(uploaded_poster_img,
                                                 'static/imgs/movie_pics')
             movie.poster = stored_poster_fn
+
         db.session.commit()
         flash(f'You have successfully updated the movie "{movie.title}"!',
               category='success')
@@ -113,12 +138,18 @@ def update_movie(movie_id, genre_id):
 
 @app.route('/genre/<int:genre_id>/movie/<int:movie_id>/delete',
            methods=['GET', 'POST'])
+@login_required
 def delete_movie(genre_id, movie_id):
     '''Delete a movie from the database.'''
 
     movie = Movie.query.get(int(movie_id))
+
+    if movie.author != current_user:
+        abort(403)
+
     movie_title = movie.title
     movie_genre = movie.genre.name
+
     if request.method == 'POST':
         db.session.delete(movie)
         db.session.commit()
@@ -128,3 +159,162 @@ def delete_movie(genre_id, movie_id):
     return render_template('delete_movie.html',
                            movie=movie,
                            title=f'Delete {movie.title}')
+
+
+# Create anti-forgery state token
+@app.route('/login')
+def google_login():
+    if current_user.is_authenticated:
+        return redirect(url_for('home'))
+
+    state = ''.join(
+        random.choice(string.ascii_uppercase + string.digits)
+        for x in range(32))
+    login_session['state'] = state
+    return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(
+                        json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code, now compatible with Python3
+    request.get_data()
+    code = request.data.decode('utf-8')
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secret.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com'
+           f'/oauth2/v1/tokeninfo?access_token={access_token}')
+    # Submit request, parse response - Python3 compatible
+    h = httplib2.Http()
+    response = h.request(url, 'GET')[1]
+    str_response = response.decode('utf-8')
+    result = json.loads(str_response)
+
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(
+            json.dumps('Current user is already connected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    # see if user exists, if it doesn't make a new one
+    user_id = get_user_id(login_session['email'])
+    if not user_id:
+        user_id = create_user(login_session)
+    login_session['user_id'] = user_id
+    user = Member.query.get(int(user_id))
+    login_user(user)
+
+    output = f'<h1>Welcome {login_session["username"]}!</h1>'
+    output += (f'<img src="{login_session["picture"]}"'
+               ' style="width:300px; height: 300px;'
+               ' border-radius: 150px;'
+               ' -webkit-border-radius: 150px;'
+               '-moz-border-radius: 150px;">')
+    flash('You are now logged in as:'
+          f' {login_session["username"].title()}!', category='success')
+    return output
+
+
+def create_user(login_session):
+    new_user = Member(username=login_session['username'], email=login_session[
+                   'email'], profile_pic=login_session['picture'])
+    db.session.add(new_user)
+    db.session.commit()
+    user = Member.query.filter_by(email=login_session['email']).first()
+    return user.id
+
+
+def get_user_id(email):
+    try:
+        user = Member.query.filter_by(email=email).first()
+        return user.id
+    except NoResultFound:
+        return None
+
+
+@app.route('/gdisconnect')
+@login_required
+def gdisconnect():
+        # Only disconnect a connected user.
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = f'https://accounts.google.com/o/oauth2/revoke?token={access_token}'
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    if result['status'] == '200':
+        # Reset the user's sesson.
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        logout_user()
+        return redirect(url_for('home'))
+    else:
+        # For whatever reason, the given token was invalid.
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
